services/auth_service/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.2
grpcio==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.2.0
protobuf==4.24.4
PyJWT==2.8.0
pytest==7.4.4
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
